+++ app/interface/src/components/Lab/Export/ExportLabButton/index.tsx
@@ -1,23 +1,14 @@
 import { saveAs } from "file-saver";
-import { useLab } from "../../hooks/useLab";
-import Button from "../Button";
+import { ButtonVariant, Lab } from "../../../../dataStructures";
+import Button from "../../../Button";
 
 type Props = {
-  variant:
-    | "primary"
-    | "secondary"
-    | "danger"
-    | "success"
-    | "primary-outline"
-    | "secondary-outline"
-    | "danger-outline"
-    | "success-outline"
-    | "primary-outline-animate";
-  children?: React.ReactNode;
+  variant: ButtonVariant;
+  lab: Lab;
+  children: React.ReactNode;
 };
 
-export default function index({ variant, children }: Props) {
-  const { data: lab } = useLab();
+export default function ExportLabButton({ variant, lab, children }: Props) {
   function handleDownload() {
     const jsonString = JSON.stringify(lab, null, 4);
     var name = "lab.json";
