+++ app/interface/src/components/Lab/Import/ImportLabButton/index.tsx
@@ -1,0 +1,5 @@
+type Props = {};
+
+export default function ImportLabButton({}: Props) {
+  return <></>;
+}
