+++ app/interface/src/pages/Templates/index.tsx
@@ -1,4 +1,5 @@
 import DeleteLabButton from "../../components/Lab/DeleteLabButton";
+import ExportLabButton from "../../components/Lab/Export/ExportLabButton";
 import LabCard from "../../components/Lab/LabCard";
 import LoadToBuilderButton from "../../components/Lab/LoadToBuilderButton";
 import TemplateCard from "../../components/TemplateCard";
@@ -74,12 +75,15 @@ function TemplateCards({ lab }: TemplateCardsProps) {
       <LabCard lab={lab}>
         <>
           <div className="flex flex-wrap justify-end gap-1">
-            <DeleteLabButton lab={lab} variant="secondary-outline">
+            <DeleteLabButton lab={lab} variant="danger-outline">
               Delete
             </DeleteLabButton>
             <LabBuilder lab={lab} variant="secondary-outline">
               Edit
             </LabBuilder>
+            <ExportLabButton lab={lab} variant="secondary-outline">
+              Export
+            </ExportLabButton>
             <LoadToBuilderButton lab={lab} variant="primary">
               Load To Builder
             </LoadToBuilderButton>
